std: Move tuple extension methods to prim
